+++ examples/distributed/participant.rs
@@ -7,8 +7,8 @@ use std::net::{SocketAddr};
 use std::collections::{HashMap};
 
 pub struct Participant {
-    event_queue: EventQueue<NetEvent>,
     network: Network,
+    event_queue: EventQueue<NetEvent>,
     name: String,
     discovery_endpoint: Endpoint,
     public_addr: SocketAddr,
@@ -17,7 +17,7 @@ pub struct Participant {
 
 impl Participant {
     pub fn new(name: &str) -> Option<Participant> {
-        let (mut network, event_queue) = Network::split();
+        let (event_queue, mut network) = Network::split();
 
         // A listener for any other participant that want to establish connection.
         // 'addr' contains the port that the OS gives for us when we put a 0.